# Conflicts:
#	.idea/sonarIssues.xml
#	.idea/sonarlint/issuestore/7/7/77110284543a45bd51793617d602b8319f3a617f
#	.idea/sonarlint/issuestore/index.pb
#	.idea/workspace.xml
#	README.md
#	config/__pycache__/setting.cpython-36.pyc
#	config/path.yaml
#	config/setting.py
#	lib/public/HtmlReport.py
#	lib/public/__pycache__/Recursion.cpython-36.pyc
#	lib/public/__pycache__/case_manager.cpython-36.pyc
#	lib/public/__pycache__/http_keywords.cpython-36.pyc
#	lib/public/__pycache__/load_cases.cpython-36.pyc
#	lib/public/__pycache__/wraps.cpython-36.pyc
#	lib/public/http_keywords.py
#	lib/public/relevance.py
#	lib/public/text_similarity_comparison.py
#	lib/public/wraps.py
#	lib/utils/__pycache__/use_SqlServer.cpython-36.pyc
#	update_log.md

+++ run.py
@@ -4,11 +4,12 @@ import click
 import unittest
 import pysnooper
 from config import setting
+from lib.utils import time_util
 from lib.utils import email
 from lib.public.case_manager import TestContainer
-from lib.public.HtmlReport import Report
 from lib.utils.analyze_log import WeChatAlarm
 from lib.utils.random_data import RandomData
+from lib.public.HtmlReport import Report
 
 
 @click.command()
@@ -19,12 +20,20 @@ from lib.utils.random_data import RandomData
     default=setting.REPORT,
     help="generator report in path")
 @pysnooper.snoop()
-def run(cases=setting.TEST_CASES, pattern='*.py', report=setting.REPORT):
+def run(cases=setting.TEST_CASES, pattern='*.py', report=setting.REPORT) -> None:
+    r"""运行测试用例主入口
 
+    :param cases: 运行的测试用例文件路径, str object.
+    :param pattern: 匹配运行的测试用例文件, str object.
+    :param report: 生成的测试报告路径, str object.
+    """
+
+    # 生成随机测试数据
     RandomData.create_random_test_data()
 
+    # 加载&运行测试用例
     test_suite = unittest.defaultTestLoader.discover(cases, pattern)
-    result = Report(test_suite, retry=3)
+    result = Report(test_suite)
     result.report(filename='Report', description='Report', log_path=report)
 
     # 临时文件回溯
@@ -33,13 +42,13 @@ def run(cases=setting.TEST_CASES, pattern='*.py', report=setting.REPORT):
             for path in filepath:
                 try:
                     for files in os.listdir(path):
-                        filename = path + files
-                        if os.path.isfile(filename):
+                        filename, pattern_back_date = path + files, time_util.timestamp('format_day')
+                        if os.path.isfile(filename) and pattern_back_date not in files.split('.'):
                             os.remove(filename)
                 except PermissionError:
                     pass
 
-    back_track_files_path = [cases, setting.Recording, setting.WORK_FLOW]
+    back_track_files_path = [cases, setting.LOG]
     files_backtrack(back_track_files_path)
 
     # 发送邮件
